Fixing occasional wrong layout.
Sometimes contentOffset of containerView, that is used for positioning of subviews during layout is not final and may change after contentSize change. This can happen when containerView is scrolled to bottom and some cells are deleted from tableView, that is located at the bottom. This triggers layoutSubviews during which contentOffset contains not updated value.