Add another demo view with subviews using auto layout
It works not quite correctly. Also, we use a hack in `ChildViewUsingAutolayout` to notify `OLEContainerScrollView` that it should update its layout when a subview's optimum size changes. This is clearly not a good solution. If we want `OLEContainerScrollView` to work properly with subviews using auto layout, we probably have to change it completely to use constraints for everything.